Revert earlier bugfix: wallmax is right!
The issue arose from extra column ('on') in GPTL summary_stats file.
We added that column to parserConfigs in model_timing.py to handle new
model outputs.

Unfortunately, when we uploaded old (DECK, MMF, high-res) experiments
from 2018-early 2019, this caused mismatched fields in parsing.

To fix that, I reverted parserConfig to old (without 'on') and uploaded
the older experiments. Anyway, that meant we had the right field in this
file all along.